+++ drift_dev/lib/src/services/schema/schema_isolate.dart
@@ -31,6 +31,7 @@ class SchemaIsolate {
     final imports = LibraryImportManager();
     final writer = Writer(
       DriftOptions.fromJson({
+        ...options.options.toJson(),
         'generate_manager': false,
         'skip_verification_code': true,
         'data_class_to_companions': false,
@@ -144,11 +145,13 @@ class SchemaIsolate {
   }
 
   static Future<List<CreateStatement>> collectStatements({
+    required DriftOptions options,
     required List<DriftElement> allElements,
     required List<DriftSchemaElement> elementFilter,
     File? dumpStartupCode,
   }) async {
     final result = await _startAndRun((
+      options: options,
       dialect: null,
       elements: allElements,
       dumpStartupCode: dumpStartupCode,
@@ -174,6 +177,7 @@ class SchemaIsolate {
 }
 
 typedef SchemaIsolateOptions = ({
+  DriftOptions options,
   List<DriftElement> elements,
   SqlDialect? dialect,
   File? dumpStartupCode,